Rename `clients service` to `things service`
