+++ dashflux/src/app/app.module.ts
@@ -23,10 +23,10 @@ import { MaterialModule } from './core/material/material.module';
 import { AuthenticationService } from './core/services/auth/authentication.service';
 import { TokenStorage } from './core/services/auth/token-storage.service';
 import { ChannelsService } from './core/services/channels/channels.service';
-import { ThingsService } from './core/services/clients/clients.service';
+import { ThingsService } from './core/services/things/things.service';
 import { MockAuthService } from './core/services/mock-auth.service';
 import { MockChannelsService } from './core/services/mock-channels.service';
-import { MockClientsService } from './core/services/mock-clients.service';
+import { MockClientsService } from './core/services/mock-things.service';
 import { ChannelsStore } from './core/store/channels.store';
 import { ThingsStore } from './core/store/things.store';
 import { UiStore } from './core/store/ui.store';
