+++ dashflux/src/app/components/channels/channels.component.spec.ts
@@ -9,7 +9,7 @@ import { MaterialModule } from '../../core/material/material.module';
 import { AuthenticationService } from '../../core/services/auth/authentication.service';
 import { TokenStorage } from '../../core/services/auth/token-storage.service';
 import { ChannelsService } from '../../core/services/channels/channels.service';
-import { ThingsService } from '../../core/services/clients/clients.service';
+import { ThingsService } from '../../core/services/things/things.service';
 import { ChannelsComponent } from './channels.component';
 import { ThingsStore } from '../../core/store/things.store';
 import { UiStore } from '../../core/store/ui.store';
