+++ dashflux/src/app/components/things/things.component.spec.ts
@@ -9,7 +9,7 @@ import { MaterialModule } from '../../core/material/material.module';
 import { AuthenticationService } from '../../core/services/auth/authentication.service';
 import { TokenStorage } from '../../core/services/auth/token-storage.service';
 import { ChannelsService } from '../../core/services/channels/channels.service';
-import { ThingsService } from '../../core/services/clients/clients.service';
+import { ThingsService } from '../../core/services/things/things.service';
 import { ThingsComponent } from './things.component';
 import { UiStore } from '../../core/store/ui.store';
 import { ThingsStore } from '../../core/store/things.store';
