+++ dashflux/src/app/core/store/ui.store.spec.ts
@@ -6,7 +6,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { AuthenticationService } from '../services/auth/authentication.service';
 import { TokenStorage } from '../services/auth/token-storage.service';
 import { ChannelsService } from '../services/channels/channels.service';
-import { ThingsService } from '../services/clients/clients.service';
+import { ThingsService } from '../services/things/things.service';
 import { UiStore } from './ui.store';
 
 describe('State', () => {
