+++ dashflux/src/environments/environment.prod.ts
@@ -2,6 +2,6 @@ export const environment = {
   production: true,
   signupUrl: '/users',
   loginUrl: '/tokens',
-  clientsUrl: '/clients',
+  thingsUrl: '/things',
   channelsUrl: '/channels'
 };
